pyproject.toml
[project]
name = "asl-detection"
version = "0.1.0"
description = "This"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
]
